on ajoute les verbes http get,post,put,delete

+++ src/main/java/com/service/security/AnimalR.java
@@ -1,7 +1,9 @@
 package com.service.security;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalR extends JpaRepository<Animal, Double> {
+public interface AnimalR extends JpaRepository<Animal, BigInteger> {
 
 }
